fix(musea): keep setup validation scoped
